requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
rich==13.9.4
pypdf==5.1.0
python-docx==1.1.2
beautifulsoup4==4.12.3
python-multipart==0.0.12
pytest==8.3.3
pytest-asyncio==0.24.0
